backend/src/main/kotlin/at/wrk/tafel/admin/backend/modules/distribution/DistributionService.kt
package at.wrk.tafel.admin.backend.modules.distribution

import at.wrk.tafel.admin.backend.common.auth.model.TafelJwtAuthentication
import at.wrk.tafel.admin.backend.database.entities.distribution.DistributionEntity
import at.wrk.tafel.admin.backend.database.repositories.auth.UserRepository
import at.wrk.tafel.admin.backend.database.repositories.distribution.DistributionRepository
import jakarta.persistence.EntityNotFoundException
import org.springframework.security.core.context.SecurityContextHolder
import org.springframework.stereotype.Service
import java.time.ZonedDateTime

@Service
class DistributionService(
    private val distributionRepository: DistributionRepository,
    private val userRepository: UserRepository
) {

    fun startDistribution(): DistributionEntity {
        val authenticatedUser = SecurityContextHolder.getContext().authentication as TafelJwtAuthentication

        val distribution = DistributionEntity()
        distribution.startedAt = ZonedDateTime.now()
        distribution.startedByUser = userRepository.findByUsername(authenticatedUser.username!!).orElse(null)

        return distributionRepository.save(distribution)
    }

    fun endDistribution(distributionId: Long) {
        val latestDistribution = distributionRepository.findById(distributionId)
            .orElseThrow { EntityNotFoundException("Distribution $distributionId not found!") }

        latestDistribution.endedAt = ZonedDateTime.now()

        distributionRepository.save(latestDistribution)
    }

    fun getCurrentDistribution(): DistributionEntity? {
        return distributionRepository.findFirstByEndedAtIsNullOrderByStartedAtDesc().orElse(null)
    }

}
